Completely rework trends module
The trends module is now capable of working with multiple countries.
Trends are pulled from each of the countries specified in config.woeid_array
and merged into one master list based on the worldwide trends list
retrieved via a call to trends/place with id=1.

Trends are returned as an array of objects of the form

{
	name: `trend name`,
	tweet_volume: `trend tweet volume`,
	locations: `array of country codes in which this trend is trending`
}

Additionally, the trends module is now completely based on promises rather
than taking and working with callbacks.